+++ test/foundry/mixins/Configurable.sol
@@ -17,7 +17,7 @@ abstract contract Configurable {
         address payable assetRecipient,
         CollectionPool.PoolType poolType,
         uint128 delta,
-        uint96 fee,
+        uint24 fee,
         uint128 spotPrice,
         uint256[] memory _idList,
         uint256 initialTokenBalance,
