+++ test/foundry/mixins/UsingERC20.sol
@@ -42,7 +42,7 @@ abstract contract UsingERC20 is Configurable, RouterCaller {
         address payable assetRecipient,
         CollectionPool.PoolType poolType,
         uint128 delta,
-        uint96 fee,
+        uint24 fee,
         uint128 spotPrice,
         uint256[] memory _idList,
         uint256 initialTokenBalance,
