+++ test/foundry/mixins/UsingETH.sol
@@ -37,7 +37,7 @@ abstract contract UsingETH is Configurable, RouterCaller {
         address payable assetRecipient,
         CollectionPool.PoolType poolType,
         uint128 delta,
-        uint96 fee,
+        uint24 fee,
         uint128 spotPrice,
         uint256[] memory _idList,
         uint256,
